+++ src/CanvasRenderer/index.js
@@ -2,5 +2,6 @@ require('./prepareImages');
 require('./vectorialRendering');
 require('./swfSymbolRendering');
 require('./renderImages');
+require('./renderImageToCanvas');
 
 module.exports = require('./main');